feat(dynamic_tune): improve tune diagnostics and fix matmul case

Created-by: wen_xiang_zou_admin
Commit-by: zouwenxiang
Merged-by: yiyanzhi_akane1
Description: <!--  Thanks for sending a pull request!  Here are some tips for you:

If this is your first time, please read our contributor guidelines: https://github.com/mindspore-ai/mindspore/blob/master/CONTRIBUTING.md
-->

**What type of PR is this?**
> Uncomment only one ` /kind <>` line, hit enter to put that in a new line, and remove leading whitespaces from that line:
>
/kind bug


**What does this PR do / why do we need it**:


**Which issue(s) this PR fixes**:
<!-- 
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes [#366](https://gitcode.com/mindspore/akg/issues/366) 

**Special notes for your reviewers**:



See merge request: mindspore/akg!2049

+++ akg_agents/tests/op/st/dynamic_tune/cases/original/matmul/impl.py
@@ -54,7 +54,7 @@ class ModelNew(nn.Module):
         BLOCK_M = 64
         BLOCK_N = 128
         BLOCK_K = 256
-        C = torch.empty(M, N, dtype=torch.float32, device=A.device)
+        C = torch.empty(M, N, dtype=A.dtype, device=A.device)
         matmul_kernel[(ASCEND_CUBE_CORE_NUM,)](
             A, B, C,
             M, N, K,
